+++ app/src/main/java/com/sackcentury/shinebutton/App.java
@@ -1,0 +1,12 @@
+package com.sackcentury.shinebutton;
+
+import android.app.Application;
+import com.google.android.material.color.DynamicColors;
+
+public class App extends Application {
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        DynamicColors.applyToActivitiesIfAvailable(this);
+    }
+}